NEON: drop compile-time feature detection, add Clang 3.5 fix
Compile-time feature detection was not working all that well across
platforms; it failed to pick up NEON in an AArch64 environment for
instance. Make it the explicit task of the user to supply the proper
architecture flags, as is usual on ARM.

Check for the __ARM_NEON flag as well as __ARM_NEON__, because Clang 3.5
defines only the former on AArch64.