Use pointers to const instead of const pointers to const
Declaring a `const char *const ptr` means you are declaring a pointer to
"readonly" memory, whose own address value is also immutable. The last
part is not relevant to library users, so remove it.

We want to communicate to the library user the important fact that the
inputs are never changed by a function, because the compiler could use
that information to do some optimization. But the library user doesn't
care whether we ever reassign the pointer internally to a different
location, because that's just an implementation detail.