lib: use proper unsigned type for walking string
Use a pointer to unsigned char as the type of variable to walk the
string and write the output. Important since we know that character
values can be over 127, and we use the value as an index into a 256-part
array during decoding. After this change, we can also remove some ugly
typecasts.