Switch to C99 (from C89).
This has been in the works for some time. It hasn't been possible to
compile this library with '-std=c89' for a while because the ARM header
files on Clang use C99 constructs. We were also using C99 features like
'stdint.h', so we were C99 in all but name.

We now need C99 because we want to use string literals longer than 512
characters, we want to use uint64_t, and we want to use inline
functions.