AVX2, SSSE3: break up long 'or' expressions
A string of five chained 'or' operations turns into a long and messy
statement, so break them up using a temporary. The code generated should
be identical to the old situation.

Also cast the return values of the `movemask` intrinsics and the
constant reference values to the correct vector width (16-bit or
32-bit), just to be explicit.