OpenMP: conditionally include into plain functions
Remove lib_plain.c and merge the code back into lib.c. OpenMP functions
are called from the plain codec if OpenMP is enabled and the data size
is larger than the threshold.

This makes the code a bit cleaner since the body of the plain codecs
does not have to be duiplicated inside the OpenMP ones.