OpenMP decode: merge return value from all threads properly
Sum the return values from the separate threads to catch decode errors
in individual threads. If the sum is `-num_threads', then all threads
returned -1, meaning that the requested codec is not available. If the
sum is not `num_threads', then at least one of the threads hit a
decoding error.