Plain decoding optimization
This now uses a modified implementation of Nick Galbreath (@client9)
algorithm.
Modifications of the original algorithm include:
- Unaligned access is now optional (macro to enable this)
- Invalid characters are mapped to 0xffffffff instead of 0x01ffffff.
This removes the need to shift by 8 the returned value of the tables on
big-endian architecture with fast unaligned access at the expense of a
different check for invalid character depending on endianness

Speed-up using clang-800.0.42.1 on i7-4870HQ @ 2.5 GHz or iPhone SE:
- x86_64: +97%
- i386: +60%
- arm64: +0%
- armv7: +24%

As a side note, it seems that the iPhone SE processor (Apple A9) has
fast unaligned access which gives even higher speed-up with
corresponding macro set to 1 (+23% / +56%). This may not be the case
for some other ARM processors.