#56: generic codecs: use memcpy for memory loads/stores
Use the `memcpy(3)' function to load and store multi-byte values to
potentially unaligned memory addresses. This function is defined in the
C99 standard, so it is portable across all standards-compliant
platforms.

Because addresses and sizes are fixed at compile time, any good compiler
will optimize the memcpy call to a single move instruction on platforms
which allow unaligned memory access, or to a series of bytewise
loads/stores on the remainder, which is as good as what we could do
manually.