NEON64: load 64-byte tables using inline assembly
Add an inline assembly routine on supported compilers to efficiently
load a 64-byte table. This replaces the previous crappy version that
works around a GCC limitation.

The limitation is that GCC will not allocate a contiguous set of
registers for a `uint8x16x4_t', so it cannot support the `vld1q_u8_x4'
intrinsic. This is worked around by replacing it with inefficient single
loads.

If we can use inline assembly (which on GCC we can), we can do better by
forcing the compiler to allocate contiguous registers and issuing the
load instruction manually.