NEON64: enc: convert full encoding loop to inline assembly
Convert the full encoding loop to an inline assembly implementation for
compilers that support inline assembly.

The motivation for this change is issue #96: when optimization is turned
off on recent versions of clang, the encoding table is sometimes not
loaded into sequential registers. This happens despite taking pains to
ensure that the compiler uses an explicit set of registers for the load
(v8..v11).

This leaves us with not much options beside rewriting the full encoding
loop in inline assembly. Only then can we be absolutely certain that the
right registers are used. Thankfully, AArch64 assembly is not very
difficult to write by hand.

In making this change, we optimize the unrolled loops for rounds >= 8 by
interleaving memory operations (loads, stores) with data operations
(arithmetic, table lookups). Splitting these two classes of instructions
avoids pipeline stalls and data dependencies. The current loop iteration
also prefetches the data needed in the next iteration.

To allow that without duplicating massive amounts of code, we abstract
the various assembly blocks into preprocessor macros and instantiate
them as needed. This mixing of the preprocessor with inline assembly is
perhaps a bit gnarly, but I think the usage is simple enough that the
advantages (code reuse) outweigh the disadvantages.

Code was tested on a Debian VM running under QEMU. Unfortunately,
testing in a VM does not let us measure the actual performance impact.