#104: AVX2: enc: add an inline assembly implementation
Add a hand-optimized inline assembly implementation of the AVX2 encoder.
This implementation uses the same algorithm as the regular AVX2 encoder,
but sets up a carefully pipelined codepath to separate memory operations
from logic operations. Loops are unrolled, and three rounds are
processed in parallel by interweaving the operations for each row.

Tests show a dramatic performance increase in the order of 20 percent or
more, depending on circumstances such as CPU type and data (block) size.

Resolves #104.