#108: AVX: enc: add an inline assembly implementation
Issue #104 introduced an inline assembly version of the AVX2 base64
encoder. It turns out that we can reuse this implementation almost
verbatim for AVX if we replace the 256-bit AVX2 registers with 128-bit
AVX registers. Some small changes are needed, such as changing the
stride widths and removing the special handling of the first round, but
overall the code remains the same.

Tested with the Intel SDE instruction set emulator running in Sandy
Bridge mode (with AVX, but no AVX2).

Resolves #108.