#111: bugfix: fix out-of-bounds read in inline asm encoders
Fix an out-of-bounds read in the inline assembly encoder
implementations.

Rounds are 12 bytes in size, but reads are done in blocks of 16 bytes.
To ensure that there is always enough space to read those 16 bytes, we
need to "reserve" four bytes of input by subtracting those bytes from
the input buffer.

Resolves #111.