#118: enc: asm: add memory and flags as clobbers
Add the memory and flags ("cc") as clobbers to all assembly encoder
implementations. This informs the compiler that the assembly will access
an unspecified amount of memory and will modify the flags. The compiler
will not use any dangling references to memory or keep a cached copy of
the flags register.

In tests, the compilers seem to output the exact same code as before, so
this should have no functional impact beyond ensuring correctness.

Resolves #118.