Detect word size for Windows properly
In Windows SDK the _INTEGRAL_MAX_BITS is always 64 even when building for
x86_32 targets, also _INTEGRAL_MAX_BITS means the max bits of integer, which
does not necessarily equal to the word size, so it should be used at all.

According to MSDN:
_WIN32 Defined as 1 when the compilation target is 32-bit ARM, 64-bit ARM, x86,
or x64. Otherwise, undefined.
_WIN64 Defined as 1 when the compilation target is 64-bit ARM or x64. Otherwise,
undefined.
So we can use them as a more reliable way for word size on Windows.