#125: fix: Detect platform word size based on `SIZE_MAX`
The native word size on a platform matches the size of `size_t` with a
few rare exceptions like the x32 ABI on linux. Therefore we can derive a
general cross platform value for `BASE64_WORDSIZE` based on `SIZE_MAX`
which the C99 standard guarantees to be available and well defined.

This fixes the misdection of x86/arm32 win32 as a 64bit platform.

See also https://en.cppreference.com/w/c/types/limits

Resolves #123.
Resolves #125.